added dithering for more spells

+++ src/game/Objects/Player.h
@@ -2486,8 +2486,8 @@ template <class T> T Player::ApplySpellMod(uint32 spellId, SpellModOp op, T &bas
         }
     }
 
-    float diff = (float)basevalue*(float)totalpct/100.0f + (float)totalflat;
-    basevalue = T((float)basevalue + diff);
+    float diff = basevalue*totalpct/100.0f + totalflat;
+    basevalue = T(basevalue + diff);
     return T(diff);
 }
